+++ tests/Fixture/Def/Concrete.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rasuvaeff\Understudy\Tests\Fixture\Def;
+
+class Concrete
+{
+    public function value(): int
+    {
+        return 1;
+    }
+}