+++ tests/Fixture/Def/Timestamped.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rasuvaeff\Understudy\Tests\Fixture\Def;
+
+interface Timestamped
+{
+    public function at(): int;
+}